refactor(grid): move the figure_grid function and class to its own module
