+++ src/plastik/__init__.py
@@ -4,6 +4,7 @@ from importlib.metadata import version
 
 from plastik import colors
 from plastik.axes import *  # noqa:F401,F403
+from plastik.grid import *  # noqa:F401,F403
 from plastik.legends import *  # noqa:F401,F403
 from plastik.percentiles import percentiles
 from plastik.ridge import *  # noqa:F401,F403
